backend/backendhw1/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
python-dotenv
python-jose
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
